refactor code to simplify main.go files in inventory service and order management service
