+++ clients/inventory/client.go
@@ -8,12 +8,12 @@ import (
 	"google.golang.org/grpc/credentials/insecure"
 )
 
-// NewInventoryClient creates a new OrderManagementServiceClient.
+// NewInventoryClient creates a new InventoryServiceClient.
 func NewInventoryClient(conn *grpc.ClientConn) inventory_v1_pb.InventoryServiceClient {
 	return inventory_v1_pb.NewInventoryServiceClient(conn)
 }
 
-// CreateInventoryClient initializes and returns an OrderManagerClient.
+// CreateInventoryClient initializes and returns an InventoryServiceClient.
 func CreateInventoryClient(host string, port int) (inventory_v1_pb.InventoryServiceClient, error) {
 	address := host + ":" + fmt.Sprintf("%d", port)
 	conn, err := grpc.NewClient(address, grpc.WithTransportCredentials(insecure.NewCredentials()))
