+++ inventory_service/grpc/inventory_service.go
@@ -1,4 +1,4 @@
-package main
+package grpc
 
 import (
 	"context"
