+++ inventory_service/main.go
@@ -1,31 +1,19 @@
 package main
 
 import (
-	"fmt"
 	"log"
-	"net"
 	"os"
 
-	inventory_v1_pb "github.com/mike_jacks/pizza_co/inventory_service/ports/grpc/v1"
-	"github.com/mike_jacks/pizza_co/config"
-	"google.golang.org/grpc"
-	"google.golang.org/grpc/reflection"
+	"github.com/mike_jacks/pizza_co/inventory_service/grpc"
 )
 
 func main() {
 
-	lis, err := net.Listen("tcp", fmt.Sprintf(":%d", config.InventoryServerPort))
-	if err != nil {
-		log.Fatalf("failed to listen: %v", err)
-		os.Stdout.Sync()
-	}
-	grpcServer := grpc.NewServer()
-	inventory_v1_pb.RegisterInventoryServiceServer(grpcServer, &inventoryServer{})
-	reflection.Register(grpcServer)
+	server := grpc.NewServer()
 
-	log.Printf("Inventory servers is running on port :%d...", config.InventoryServerPort)
-	if err := grpcServer.Serve(lis); err != nil {
-		log.Fatalf("failed to serve: %v", err)
+	if err := server.Start(); err != nil {
+		log.Fatalf("Failed to start gRPC server: %v", err)
 		os.Stdout.Sync()
 	}
+
 }
