+++ order_management_service/grpc/order_management_service.go
@@ -1,4 +1,4 @@
-package main
+package grpc;
 
 import (
 	"context"
